requirements_user.txt
PyMySQL
peewee
passlib
grpcio
grpcio-tools
protobuf
loguru
python-consul2